Remove unused and untested DraggableHandler code
This CL removes code that is used when previewing a scene transition. I
am not sure why it was there but removing it does not make any test fail
and the preview case of the demo still works as expected, so let's
remove it.

Bug: 353679003
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: com.android.systemui.scene_container
Change-Id: If0ac070cd22e10012e4fd919a0e4cfd20bc07e7b