Rename IME Switcher menu button to Settings
As the "language" aspect of the button was somewhat implied with the
list of keyboard languages visible, and also as the button with the
"Language settings" text would not fit correctly in some locales,
this renames the button text to "Settings", but keeps the
contentDescription used for accessibility to "Language Settings".

Also makes clear that the Intent used by the button will try to open the
LanguageSettingsActivity, and otherwise fall back to the general
SettingActivity.

Flag: android.view.inputmethod.ime_switcher_revamp
Test: atest InputMethodManagerTest#testInputMethodPickerOpenLanguageSettings
Bug: 358249049
Change-Id: I34413b6ad3ad823f73d02d80767cfe8312b42ff2