Fix parameter for Gatekeeper re-enroll in SyntheticPasswordManager
Gatekeeper's re-enroll feature requires the enroll() API to be
called with currentPasswordHandle, currentPassword, and newPassword.

Currently, the currentPassword parameter in
SyntheticPasswordManager is incorrectly set to spHandle instead of
gatekeeperPassword. This causes the re-enroll feature to fail, as
Gatekeeper receives a password handle instead of the expected
password, leading to a mismatch during the re-enrollment process.

Change-Id: Iad84a23e72940ad04aa9d663e871bf2cef0ce1cf