Remove @DisabledOnNonRavenwood
I added it "just in case we need it", but it was a weird annotation
because it'll only affect device side behavior yet it'll require a
RavenwoodRule to take effect.

With Ravenizer, it'd be even more weird because now all other annotations
would be handled by the implicit test runner, but @DisabledOnNonRavenwood
would still need an explicit @DisabledOnNonRavenwood.

So let's just remove it. It's not used anywhere yet.

With Ravenizer, we could do it differently -- for example, we could
add an @UnIgnoreOnRavenwood, which is to be used with @Ignore,
and have Ravenizer remove the @Ignore.

Flag: EXEMPT host test change only
Bug: 292141694
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh
Change-Id: Icfd5e33f07e217a95012164dc8fa28f7d0ba6daa