Move desktop apps to front on fullscreen launch
Moving the home task to back caused it to be behind the fullscreen tasks
as well. This resulted in back navigation ending up on a fullscreen task
rather than the home screen. With this change when the app intent is
moving home task behind the newly launched app, we bring home to front
first then move all the desktop tasks on top of it.

Test: atest DesktopTasksController
Bug: 358552703
Flag: EXEMPT Bugfix
Change-Id: Ia3419f2924e05f8f4cb607678f74c048bb888377