Update bounds of shade view on config change
If the shade window is not visible, the bounds will not update until
it becomes visible. Touches that should invoke shade expansion but
are not within those incorrect bounds (because the shape of the shade
window remains portrait after flipping to landscape) will be dropped,
causing the shade expansion to fail silently. Since the shade doesn't
open, it doesn't become visible, and the bounds will never update.
Therefore, we must detect the incorrect bounds here and force the
update so that touches are routed correctly.

Test: added new tests and ran existing tests
Test: manually verified that the fix only triggers when needed
Flag: com.android.systemui.scene_container
Fixes: 360051920
Change-Id: Idb85c281de6c35821cd62d15cdcb9ba52362dde5