Keep relaunched freeform tasks in freeform if default desktop
This CL prevents a relaunched freeform task from being in fullscreen if
the default-desktop flag is enabled.
Currently, when a freeform task is relaunched out of a desktop session,
it’s moved to fullscreen. But with default-desktop, we should keep it in
freeform.

Bug: 361419732
Flag: com.android.window.flags.enter_desktop_by_default_on_freeform_displays
Test: DesktopTasksControllerTest
Change-Id: I06ec124ffc3c27eb8bb62e07271f37b2b74f26ac