Add ability to make transitions not replaceable
On the lockscreen, if the user swipes down to expand the shade then
swipes up past the origin of the swipe without lifting their finger,
it will not only collapse the shade but also unlock the phone. This
happens because when the finger passes the origin point of the swipe,
it switches from the down action (i.e. shade expand) to the up action
(i.e. unlock the phone). This CL adds the ability to mark some swipe
action results as irreversable and sets that parameter for the
lockscreen swipe down actions.

Fixes: 359876433
Test: added new tests
Test: manually verified
Flag: com.android.systemui.scene_container
Change-Id: If841ed9b1b4484d767db114b16226c4411cf7f16