Allow sqlite in isolated processes
The SQLite framework always tests the compatibility WAL flag.
However, the flag is not accessible to isolated processes, and testing
the flag generates an exception.

This change makes the code ignore exceptions when fetching the flag,
just as it ignores a null ActivityThread when fetching the flag.

Flag: EXEMPT bug-fix
Bug: 357905837
Test: atest
 * FrameworksCoreTests:android.database
 * CtsDatabaseTestCases
Change-Id: I577dabcd409b12d60ec7e24d1dbdfc4608aa5d57