Prevent pip launch during drag to desktop.
The launch of the home intent used at the start of drag to desktop
results in the dragged app attempting to invoke pip if the task is
dragged to split and if pip is applicable. These simultaneous
transitions leave the app in a broken state.

This CL avoids this by explicitly preventing pip in the first
WindowContainerTransaction of drag to desktop.

Fix: 356297301
Test: Manual, play a video in Youtube and drag to split while playing
Flag: EXEMPT bugfix
Change-Id: I56ae52d96fbeaf123e519372ba663f424b5981a3