Fullscreen Magnification updates pointer icon on non-transient change
Changing pointer icon scale involves resource loading, which is
expensive. Thus, we want to change the size of pointer icon only when
the change is not transient, i.e., not during animation, user's gesture,
moving a slider bar in settings panel, etc.

In order to update it, this change adds `isScaleTransient` parameter to
all the entrance point of changing fullscreen magnification scale so
that when the change is actually applied, we can tell whether we should
update a pointer icon size.

The change itself is big, but most of them are mechanical change of
method signatures and the behavior is controlled behind the aconfig
flag.

Bug: 355734856
Test: Enable flag, and changing scale with slider or gesture will change pointer icons.
Test: atest com.android.server.accessibility.magnification
Flag: com.android.server.accessibility.magnification_enlarge_pointer

Change-Id: Ibd8c0aa05c047c3977f8d2e7fc5d50f2193705fa