Allow loading of host runtime for Windows
This makes the loading of host runtime dependent on a Java system
property instead of simply ignoring it for Windows. By default, this
won't be loaded (keeping the current behaviour working). But setting
use_base_native_hostruntime to true will allow the host runtime to be
created when libandroid_runtime is loaded.

Flag: NONE host-only change
Bug: 340885449
Test: N/A
Change-Id: I308e43e9edbd4a468caa9a0c50b31c492d89417f