Remove redundant flows from SharedNotificationContainerInteractor
There is already an existing flow for configuration changes in
`configurationInteractor`, and it's a better practice to depend on this
interactor than the underlying repository anyway.

The other flow, `isSplitShadeEnabled`, is redundant since its usages
have been removed in ag/29864236.

This is part of a refactoring that will improve NSSL positioning to
support Dual Shade and Connected Displays.

Bug: 338033836
Flag: com.android.systemui.scene_container
Test: Existing unit tests still pass.
Change-Id: I736234dac0692468980ea126c328cbbc479fc099