Support Brightness mirror
Convert some viewmodels to activatables.

In particular, viewmodels that
use the BrightnessViewModel can specify if they want it to be mirrored
(in which case BrightnessMirrorViewModel will be notified) and the
composable will be put in an overlay while it's being dragged.

Bug: 374333334
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Test: manual
Test: atest BrightnessSliderViewModelTest
Test: atest SwipeBrightness
Test: atest NotificationShadeWindowViewControllerTest
Change-Id: I929cbe2ce4ba519088a0e712438a8d6359e0e7f0